Fix a small problem

+++ player.py
@@ -77,7 +77,8 @@ def worker():
         process = None
         if q:
             if REMOVE_AFTER_PLAYING:
-                os.remove("downloads/" + item["file"])
+                if q_list[0]["file"] != item["file"] and q_list[1]["file"] != item["file"]:
+                    os.remove("downloads/" + item["file"])
             q.task_done()
